feat: move constructor layout code to Lean and fix asymmetries
This PR moves the constructor layout code from C++ to Lean. When
writing the new compiler, we just reused the existing C++ code,
even though it was a bit inconvenient, because we wanted to
ensure that constructor layout always matched the existing
compiler.

This fixes #2589. Originally, I wanted to port the code to Lean
without fixing this issue, and then fix it afterwards. However,
I found that it was less code to just implement it "correctly"
and get this fix as a consequence than it would be to emulate
the bugs of the existing C++ implementation.